requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
drf-yasg==1.21.7
setuptools==75.1.0
